Move app code to single file
